Refactored service classes for Upper, Trig, Stock and PistolGrip classes.

+++ src/main/java/com/awoodcock/armory/service/ar/PistolGripService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -28,8 +29,9 @@ public class PistolGripService {
     }
 
     // READ
-    public Optional<PistolGrip> getPistolGripById (int id) {
-        return pistolGripRepository.findById(id);
+    public PistolGrip getPistolGripById (int id) {
+        Optional<PistolGrip> pistolGrip = pistolGripRepository.findById(id);
+        return pistolGrip.orElse(null);
     }
 
     public List<PistolGrip> getAllPistolGrips () {
@@ -37,9 +39,58 @@ public class PistolGripService {
     }
 
     // UPDATE
+    public PistolGrip updatePistolGripById(int id, PistolGrip pistolGrip) {
+        Optional<PistolGrip> pistolGripOptional = pistolGripRepository.findById(id);
+
+        if (pistolGripOptional.isPresent()) {
+            PistolGrip originalPistolGrip = pistolGripOptional.get();
+
+            // name
+            if (Objects.nonNull(pistolGrip.getName()) && !"".equalsIgnoreCase(pistolGrip.getName())) {
+                originalPistolGrip.setName(pistolGrip.getName());
+            }
+
+            // description
+            if (Objects.nonNull(pistolGrip.getDescription()) && !"".equalsIgnoreCase(pistolGrip.getName())) {
+                originalPistolGrip.setDescription(pistolGrip.getDescription());
+            }
+
+            // price
+            if (Objects.nonNull(pistolGrip.getPrice()) && pistolGrip.getPrice() != 0) {
+                originalPistolGrip.setPrice(pistolGrip.getPrice());
+            }
+
+            // weight
+            if (Objects.nonNull(pistolGrip.getWeight()) && pistolGrip.getWeight() != 0) {
+                originalPistolGrip.setWeight(pistolGrip.getWeight());
+            }
+
+            // texture
+            if (Objects.nonNull(pistolGrip.getTexture()) && !"".equalsIgnoreCase(pistolGrip.getTexture())) {
+                originalPistolGrip.setTexture(pistolGrip.getTexture());
+            }
+
+            // color
+            if (Objects.nonNull(pistolGrip.getColor()) && !"".equalsIgnoreCase(pistolGrip.getColor())) {
+                originalPistolGrip.setColor(pistolGrip.getColor());
+            }
+
+            // hasCompartment
+            if (Objects.nonNull(pistolGrip.hasCompartment())) {
+                originalPistolGrip.setCompartment(pistolGrip.hasCompartment());
+            }
+
+            return pistolGripRepository.save(originalPistolGrip);
+        }
+        return null;
+    }
 
     // DELETE
-    public void deletePistolGrip (int id) {
-        pistolGripRepository.deleteById(id);
+    public String deletePistolGrip (int id) {
+        if (pistolGripRepository.findById(id).isPresent()) {
+            pistolGripRepository.deleteById(id);
+            return "Pistol Grip successfully deleted.";
+        }
+        return "Pistol Grip not found within database.";
     }
 }
