+++ src/main/java/com/awoodcock/armory/service/ar/StockService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -28,17 +29,62 @@ public class StockService {
     }
 
     // READ
-    public Optional<Stock> getStockById (int id) {
-        return stockRepository.findById(id);
+    public Stock getStockById (int id) {
+        Optional <Stock> stock = stockRepository.findById(id);
+        return stock.orElse(null);
     }
 
     public List<Stock> getAllStocks () {
         return stockRepository.findAll();
     }
     // UPDATE
+    public Stock updateStockById(int id, Stock stock) {
+        Optional<Stock> stockOptional = stockRepository.findById(id);
+
+        if (stockOptional.isPresent()) {
+            Stock originalStock = stockOptional.get();
+
+            // name
+            if (Objects.nonNull(stock.getName()) && !"".equalsIgnoreCase(stock.getName())) {
+                originalStock.setName(stock.getName());
+            }
+
+            // description
+            if (Objects.nonNull(stock.getDescription()) && !"".equalsIgnoreCase(stock.getName())) {
+                originalStock.setDescription(stock.getDescription());
+            }
+
+            // price
+            if (Objects.nonNull(stock.getPrice()) && stock.getPrice() != 0) {
+                originalStock.setPrice(stock.getPrice());
+            }
+
+            // weight
+            if (Objects.nonNull(stock.getWeight()) && stock.getWeight() != 0) {
+                originalStock.setWeight(stock.getWeight());
+            }
+
+            // color
+            if (Objects.nonNull(stock.getColor()) && !"".equalsIgnoreCase(stock.getColor())) {
+                originalStock.setColor(stock.getColor());
+            }
+
+            // isAdjustable
+            if (Objects.nonNull(stock.isAdjustable())) {
+                originalStock.setAdjustable(stock.isAdjustable());
+            }
+
+            return stockRepository.save(originalStock);
+        }
+        return null;
+    }
 
     // DELETE
-    public void deleteStock(int id) {
-        stockRepository.deleteById(id);
+    public String deleteStock(int id) {
+        if (stockRepository.findById(id).isPresent()) {
+            stockRepository.deleteById(id);
+            return "Stock successfully deleted from database.";
+        }
+        return "Stock not found within database.";
     }
 }
