+++ src/main/java/com/awoodcock/armory/service/ar/TrigService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -28,8 +29,9 @@ public class TrigService {
     }
 
     // READ
-    public Optional<Trig> getTrigById (int id) {
-        return trigRepository.findById(id);
+    public Trig getTrigById (int id) {
+        Optional<Trig> trig = trigRepository.findById(id);
+        return trig.orElse(null);
     }
 
     public List<Trig> getAllTrigs () {
@@ -37,9 +39,47 @@ public class TrigService {
     }
 
     // UPDATE
+    public Trig updateTrigById(int id, Trig trig) {
+        Optional<Trig> trigOptional = trigRepository.findById(id);
+
+        if (trigOptional.isPresent()) {
+            Trig originalTrig = trigOptional.get();
+
+            // name
+            if (Objects.nonNull(trig.getName()) && !"".equalsIgnoreCase(trig.getName())) {
+                originalTrig.setName(trig.getName());
+            }
+
+            // description
+            if (Objects.nonNull(trig.getDescription()) && !"".equalsIgnoreCase(trig.getName())) {
+                originalTrig.setDescription(trig.getDescription());
+            }
+
+            // price
+            if (Objects.nonNull(trig.getPrice()) && trig.getPrice() != 0) {
+                originalTrig.setPrice(trig.getPrice());
+            }
+
+            // weight
+            if (Objects.nonNull(trig.getWeight()) && trig.getWeight() != 0) {
+                originalTrig.setWeight(trig.getWeight());
+            }
+
+            if (Objects.nonNull(trig.getPullWeight()) && trig.getPullWeight() != 0) {
+                originalTrig.setPullWeight(trig.getPullWeight());
+            }
+
+            return trigRepository.save(originalTrig);
+        }
+        return null;
+    }
 
     // DELETE
-    public void deleteTrig(int id) {
-        trigRepository.deleteById(id);
+    public String deleteTrig(int id) {
+        if (trigRepository.findById(id).isPresent()) {
+            trigRepository.deleteById(id);
+            return "Trigger successfully deleted.";
+        }
+        return "Trigger not found within database.";
     }
 }
